requirements.txt
ollama
pyaudio
pydub
ffmpeg
pyttsx3
openai-whisper
openai